requirements.txt
python-dotenv==0.21.0
redis==4.4.0
requests==2.28.1
rq==1.11.1
